Properly block ads server from https

+++ src/main/java/com/pnunes/sabproxy/AdServers.java
@@ -17,6 +17,7 @@ public class AdServers {
     private final Logger log = LoggerFactory.getLogger(this.getClass());
 
     public static String AD_SERVERS_FILE = "adservers.txt";
+    public static int AD_SERVERS_FILE_VALID_DAYS = 1;
     private static String AD_SERVERS_SOURCE_URL = "https://raw.githubusercontent.com/StevenBlack/hosts/master/alternates/fakenews-gambling/hosts";
 
     private String DOMAIN_PATTERN = "^((?!-)[A-Za-z0-9-]{1,63}(?<!-)\\.)+[A-Za-z]{2,6}$";
@@ -105,16 +106,30 @@ public class AdServers {
         return Utils.getAppSettingFolder()+"/"+AD_SERVERS_FILE;
     }
 
-    public void updateAdServersList(){
+    public void updateAdServersList(boolean forceUpdate){
         log.info("Trying to update ad servers list...");
-        try {
-            FileUtils.copyURLToFile(new URL(AD_SERVERS_SOURCE_URL), new File(getAdServersListFile()));
-        } catch (MalformedURLException e) {
-            log.warn("Can't update ads servers list from: "+AD_SERVERS_SOURCE_URL+". "+e.getMessage());
-        } catch (IOException e) {
-            log.warn("Failed to save ads servers list. "+e.getMessage());
+
+        // check if file exists
+        File adServersHostFile = new File(getAdServersListFile());
+        if(adServersHostFile.exists()){
+            // check if its older than x days
+            long timeDiff = new Date().getTime() - adServersHostFile.lastModified();
+            if (timeDiff > AD_SERVERS_FILE_VALID_DAYS * 24 * 60 * 60 * 1000 || forceUpdate) { // update file only if older than AD_SERVERS_FILE_VALID_DAYS
+                try {
+                    FileUtils.copyURLToFile(new URL(AD_SERVERS_SOURCE_URL), new File(getAdServersListFile()));
+                } catch (MalformedURLException e) {
+                    log.warn("Can't update ads servers list from: "+AD_SERVERS_SOURCE_URL+". "+e.getMessage());
+                } catch (IOException e) {
+                    log.warn("Failed to save ads servers list. "+e.getMessage());
+                }
+                log.info("Ad servers list updated.");
+            }else{
+                log.info("Ad servers list less than "+AD_SERVERS_FILE_VALID_DAYS+" day(s) old. Not updating.");
+            }
         }
-        log.info("Ad servers list updated.");
+
+
+
     }
 
 }
